Move stl_mesh out to its own file
